+++ shelter-angular/src/app/app.module.ts
@@ -2,16 +2,16 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MainModule } from './core/main/main.module';
-import { AdminModule } from './core/admin/admin.module';
+import { AdminModule } from './features/admin/admin.module';
 import { HttpClientModule } from '@angular/common/http';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
-import { HeaderComponent } from './core/common/header/header.component';
-import { FooterComponent } from './core/common/footer/footer.component';
-import { PetsModule } from './shared/pets/pets.module';
+import { HeaderComponent } from './core/header/header.component';
+import { FooterComponent } from './core/footer/footer.component';
+import { MainModule } from './features/main/main.module';
+import { PetsModule } from './features/common/pets/pets.module';
 
 @NgModule({
   declarations: [
