--- shelter-angular/src/app/core/admin/add-button/add-button.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-add-button',
-  templateUrl: './add-button.component.html',
-  styleUrls: ['./add-button.component.scss']
-})
-export class AddButtonComponent {
-  addPet() {
-    return null;
-  }
-}
