+++ shelter-angular/src/app/core/footer/footer.component.spec.ts
