+++ shelter-angular/src/app/core/footer/footer.component.ts
