+++ shelter-angular/src/app/core/header/header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../palette.scss';
+@import '../../../palette.scss';
 
 :host {
   background-size: cover;
