+++ shelter-angular/src/app/features/admin/add-pet-button/add-pet-button.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-add-pet-button',
+  templateUrl: './add-pet-button.component.html',
+  styleUrls: ['./add-pet-button.component.scss']
+})
+export class AddPetButtonComponent {
+  addPet() {
+    return null;
+  }
+}
