+++ shelter-angular/src/app/features/admin/admin.component.html
@@ -2,6 +2,6 @@
   <h3 class="admin__header">Administration <br> tool</h3>
   <div class="admin__cards">
     <app-pets></app-pets>
-    <app-add-button></app-add-button>
+    <app-add-pet-button></app-add-pet-button>
   </div>
 </div>
