+++ shelter-angular/src/app/features/admin/admin.component.scss
