+++ shelter-angular/src/app/features/admin/admin.module.ts
@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AdminRoutingModule } from './admin-routing.module';
-import { AdminComponent } from './admin.component';
-import { PetsModule } from 'src/app/shared/pets/pets.module';
-import { AddButtonComponent } from './add-button/add-button.component';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
+import { AdminComponent } from './admin.component';
+import { AddPetButtonComponent } from './add-pet-button/add-pet-button.component';
+import { PetsModule } from '../common/pets/pets.module';
 
 @NgModule({
   declarations: [
     AdminComponent,
-    AddButtonComponent,
+    AddPetButtonComponent,
   ],
   imports: [
     CommonModule,
