+++ shelter-angular/src/app/features/common/pets/interfaces/AddPetRequestBody.ts
