+++ shelter-angular/src/app/features/common/pets/interfaces/DeletePetResponse.ts
