+++ shelter-angular/src/app/features/common/pets/interfaces/Pet.ts
