+++ shelter-angular/src/app/features/common/pets/interfaces/UpdatePetRequestBody.ts
