+++ shelter-angular/src/app/features/common/pets/pet-big-card/pet-big-card.component.html
@@ -34,6 +34,8 @@
 
 </mat-dialog-content>
 <mat-dialog-actions align="end">
-  <button mat-button mat-dialog-close>Close</button>
+  <button mat-button *ngIf="isAdminRoute" (click)="editPet()">Edit</button>
+  <button mat-button *ngIf="isAdminRoute" (click)="deletePet()">Delete</button>
+  <button mat-button mat-dialog-close cdkFocusInitial>Close</button>
   <!-- <button mat-button [mat-dialog-close]="true" cdkFocusInitial>Adopt an animal</button> -->
 </mat-dialog-actions>
