+++ shelter-angular/src/app/features/common/pets/pet-big-card/pet-big-card.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../palette.scss';
+@import '../../../../../palette.scss';
 
 .dialog {
   display: flex;
