+++ shelter-angular/src/app/features/common/pets/pet-big-card/pet-big-card.component.spec.ts
