+++ shelter-angular/src/app/features/common/pets/pet-big-card/pet-big-card.component.ts
@@ -1,12 +1,14 @@
-import { Component, Inject } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-pet-big-card',
   templateUrl: './pet-big-card.component.html',
   styleUrls: ['./pet-big-card.component.scss']
 })
-export class PetBigCardComponent {
+export class PetBigCardComponent implements OnInit {
+  isAdminRoute = false;
   constructor(@Inject(MAT_DIALOG_DATA) public data: {
     name: string;
     animalType: string;
@@ -16,5 +18,18 @@ export class PetBigCardComponent {
     inoculations: string;
     diseases: string;
     parasites: string;
-  }) { }
+  },
+  private router: Router) { }
+
+  ngOnInit(): void {
+    this.isAdminRoute = (this.router.url === '/admin');
+  }
+
+  editPet() {
+    return "Edit pet method result is here";
+  }
+
+  deletePet() {
+    return "Delete pet method result is here";
+  }
 }
