+++ shelter-angular/src/app/features/common/pets/pet-small-card/pet-small-card.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../palette.scss';
+@import '../../../../../palette.scss';
 
 .pet-small-card {
   display: flex;
