+++ shelter-angular/src/app/features/common/pets/pet-small-card/pet-small-card.component.spec.ts
