+++ shelter-angular/src/app/features/common/pets/pets.component.scss
