+++ shelter-angular/src/app/features/main/about/about.component.ts
