+++ shelter-angular/src/app/features/main/intro/intro.component.spec.ts
