+++ shelter-angular/src/app/features/main/intro/intro.component.ts
