+++ shelter-angular/src/app/features/main/main.component.scss
