+++ shelter-angular/src/app/shared/constants.ts
@@ -1,6 +1,6 @@
 export const CONSTANTS = {
-  API_URL: "https://shelter-qnc2.onrender.com/pets",
-  // API_URL: "http://localhost:3000/pets",
+  // API_URL: "https://shelter-qnc2.onrender.com/pets",
+  API_URL: "http://localhost:3000/pets",
   PET_DEFAULT_NAME: "Bobik",
   PET_DEFAULT_TYPE: "Animal",
   PET_DEFAULT_BREED: "Without breed",
